requirements.txt
weaviate-client>=4.0.0
fastapi
uvicorn
sentence-transformers
openai-clip
torch
torchvision
pillow
numpy